Add base class for knex models
